backend/requirements.txt

fastapi
uvicorn[standard]
httpx
python-dotenv

# Testing & Linting
pytest
ruff
